required
оно работает

+++ src/TileMap/TileMap.java
@@ -1,6 +1,5 @@
 package TileMap;
 
-import Main.Game;
 import Main.GamePanel;
 
 import javax.imageio.ImageIO;
@@ -105,8 +104,8 @@ public class TileMap {
     }
 
     public int getTileSize() {return tileSize;}
-    public double getx() {return x;}
-    public double gety () {return y;}
+    public int getx() {return (int)x;}
+    public int gety () {return (int)y;}
     public int getWidth() {return width;}
     public int getHeight() {return height;}
